# internal/handler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Constellation Control Plane API",
    "version": "0.1.0",
    "description": "Phase 1 control-plane endpoints. Astronomer integration surfaces /api/v1/security/* alongside the standalone /api/v1/* routes."
  },
  "servers": [
    {
      "url": "http://localhost:18080"
    },
    {
      "url": "https://constellation.example.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error envelope returned by API handlers.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/vex/cyclonedx/{asset_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Get cyclonedx"
      }
    },
    "/api/v1/vex/openvex/{asset_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Get openvex"
      }
    },
...
  }
}